requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
colorama>=0.4.6
python-dotenv>=1.0.0
